\name{CTAC35}
\alias{CTAC35}
\docType{data}
\title{
CTAC 35 database.
}
\description{
A database to be used as example in the functions included on \code{InDisc} package.
}
\usage{data("CTAC35")}
\format{
  A matrix with 758 observations and 35 variables corresponding to the CTAC questionnaire (Pallero, Ferrando, & Lorenzo-Seva, 1998).
}
\details{
The CTAC questionnaire (Pallero, Ferrando, & Lorenzo-Seva, 1998) measures anxiety in situations related to visual deficit and which is intended to be used in the general adult population with severe visual impairment. The CTAC items use a 5-point Likert format.
}
\references{
Pallero, R., Ferrando, P.J., & Lorenzo-Seva, U. (1998). Questionnaire Tarragona of anxiety for blind people. In E. Sifferman, M. Williams, & B.B. Blasch, (eds.): The 9th Internacional Mobility Conference Proceedings, (pp 250-253). Atlanta: Rehabilitation Research and Development Center.
}
\examples{
data(CTAC35)
}
\keyword{datasets}
