\name{InDisc-internal}
\alias{InDisc-internal}
\alias{chis}
\alias{creanchi}
\alias{creanodos}
\alias{eap17c}
\alias{eap17g}
\alias{eap17gth}
\alias{eapth17c}
\alias{lrtseg}
\alias{lrtest}
\alias{pcong}
\alias{priorgral}
\alias{reap17c}
\alias{reap17g}
\alias{reapth17c}
\alias{reapth17g}
\alias{rlrtesg}
\alias{rlrtest}
\alias{thresholds}
\alias{verosi2}
\alias{betagradm}
\alias{eapmulc2}
\alias{eapmulc3}
\alias{eapmulc4}
\alias{eapmulg2}
\alias{eapmulg3}
\alias{eapmulg4}
\alias{eapmulthc2}
\alias{eapmulthc3}
\alias{eapmulthc4}
\alias{generbeta}
\alias{lrtestm}
\alias{pcongmul}
\alias{priormul}
\alias{reapmulc}
\alias{reapmulg}
\alias{reapmulthc}
\alias{rlrtestm}
\alias{rthres}
\alias{thres3}
\alias{thres4}
\alias{thres5}
\alias{thresdic}
\alias{size}
\alias{transpose}
\title{
InDisc-internal
}
\description{
This functions are for internal use only.
}
\usage{
chis
creanchi
creanodos
eap17c
eap17g
eap17gth
eapth17c
lrtseg
lrtest
pcong
priorgral
reap17c
reap17g
reapth17c
reapth17g
rlrtesg
rlrtest
thresholds
verosi2
size
transpose
}
\author{
Pere Joan Ferrando

David Navarro-Gonzalez
}
\keyword{ internal }
